+++ sensors/src/SensorBase.cpp
@@ -1,8 +1,9 @@
 // Copyright 2023 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
+#include <stdio.h>
+
 #include "SensorBase.hpp"
-#include "stdio.h"
 
 using namespace sensor;
 
@@ -57,9 +58,7 @@ int SensorBase::i2c_write_table(i2c_table_t table) {
   int ret = 0;
 
   for (size_t i = 0; i < table.num_lines; i++) {
-    //uint16_t address = table[i].reg_addr;
     uint16_t address = table.table[i].reg_addr;
-    //uint16_t value = table[i].reg_val;
     uint16_t value = table.table[i].reg_val;
     
     // pause if we reset the device
@@ -89,5 +88,36 @@ int SensorBase::i2c_write_table(i2c_table_t table) {
       ret |= this->i2c_write_line(address, (uint8_t)value);
     }
   }
-  return ret;
+  return ret != I2C_REGOP_SUCCESS ? -1 : 0;
+}
+
+int SensorBase::initialize() {
+  xassert(0 && "Sensor Exception: Make sure your initialize() method is implemented and called from the derived class");
+  return -1;
+}
+
+int SensorBase::stream_start() {
+  xassert(0 && "Sensor Exception: Make sure your stream_start() method is implemented and called from the derived class");
+  return -1;
+}
+
+int SensorBase::stream_stop() {
+  xassert(0 && "Sensor Exception: Make sure your stream_stop() method is implemented and called from the derived class");
+  return -1;
+}
+
+int SensorBase::set_exposure(uint32_t dBGain) {
+  (void)dBGain;
+  xassert(0 && "Sensor Exception: Make sure your set_exposure() method is implemented and called from the derived class");
+  return -1;
+}
+
+int SensorBase::configure() {
+  xassert(0 && "Sensor Exception: Make sure your configure() method is implemented and called from the derived class");
+  return -1;
+}
+
+void SensorBase::control(chanend_t c_control) {
+  (void)c_control;
+  xassert(0 && "Sensor Exception: Make sure your control() method is implemented and called from the derived class");
 }